Renaming Landing -> HomePage and routing authentication pages
