+++ frontend/src/components/HomePage.vue
